feat: further tweaks
