+++ app/core/lib/curations/commandBus.ts
@@ -1,15 +1,19 @@
 import type { CommandDefinition } from "~/types";
-import { commandMeta } from "./meta";
-import { coreActions } from "./handlers/coreActions";
 import { nounsActions } from "./handlers/nounsActions";
+import { nounsCommands } from "./commands/nounsCommands";
 
-const commandHandlers = [...coreActions, ...nounsActions];
+export const commandMeta: Pick<CommandDefinition, "id" | "meta">[] = [
+  ...nounsCommands,
+];
+
+const commandHandlers = [...nounsActions];
 
 export const commandBus: CommandDefinition[] = commandMeta.map((meta) => {
   const id = meta.id;
-  const matchedHandler = commandHandlers.find((h) => h.id === id);
 
-  if (!matchedHandler) {
+  const handler = commandHandlers.find((h) => h.id === id);
+
+  if (!handler) {
     throw new Error(`No handler found for command: ${id}`);
   }
 
@@ -23,7 +27,7 @@ export const commandBus: CommandDefinition[] = commandMeta.map((meta) => {
           typeof v === "string" ? v.trim().toLowerCase() : v,
         ])
       ) as any;
-      await matchedHandler.action(cleanArgs);
+      await handler.action(cleanArgs);
     },
     handler: async () => {
       throw new Error(`Handler not found for command: ${id}`);
