+++ app/core/lib/curations/helpers.ts
